+++ api/main.py
@@ -6,7 +6,7 @@ import logging
 from api.config import load_config
 from api.fraction_functions import mixed_number_to_fraction
 from api.data_manager import load_data, save_data
-from models.models import UserData, DebtEntry, OweRequest, SettleRequest, SetUnicodePreferenceRequest
+from models import UserData, DebtEntry, OweRequest, SettleRequest, SetUnicodePreferenceRequest
 
 # Setup 
 logging.basicConfig(level=logging.DEBUG)
