+++ models/pint_economy.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel, Field
+from typing import Dict
+
+from models.user_data import UserData
+
+class PintEconomy(BaseModel):
+    users: Dict[str, UserData] = Field(default_factory=dict)
