+++ models/set_unicode_preference_request.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class SetUnicodePreferenceRequest(BaseModel):
+    user_id: str
+    use_unicode: bool
