+++ models/settle_request.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class SettleRequest(BaseModel):
+    debtor: int
+    creditor: int
+    amount: str
