+++ models/user_debts.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel, Field
+from typing import List, Dict
+
+from models.debt_entry import DebtEntry
+
+class UserDebts(BaseModel):
+    creditors: Dict[str, List[DebtEntry]] = Field(default_factory=dict)
