+++ models/user_preferences.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class UserPreferences(BaseModel):
+    use_unicode: bool = False
